+++ mfs/context.go
@@ -6,6 +6,7 @@ import (
 	"github.com/miacio/varietas/util"
 )
 
+// Mfs
 type Mfs interface {
 	Next(...int)            // Next
 	Back(...int)            // Back
@@ -34,6 +35,7 @@ type Context struct {
 	err      error          // context error
 }
 
+// Next do method next to index[0]
 func (ctx *Context) Next(index ...int) {
 	next := 1
 	if index != nil {
@@ -42,6 +44,7 @@ func (ctx *Context) Next(index ...int) {
 	ctx.next <- ctx.Now + next
 }
 
+// Back do method back to index[0]
 func (ctx *Context) Back(index ...int) {
 	back := 1
 	if index != nil {
@@ -50,28 +53,34 @@ func (ctx *Context) Back(index ...int) {
 	ctx.next <- ctx.Now - back
 }
 
+// Stop wait the method loding other method over to Start
 func (ctx *Context) Stop() {
 	ctx.stop.Add(1)
 	ctx.stop.Wait()
 }
 
+// Start
 func (ctx *Context) Start() {
 	ctx.stop.Done()
 }
 
+// Clase
 func (ctx *Context) Close(err error) {
 	ctx.next <- -1
 	ctx.err = err
 }
 
+// Error
 func (ctx *Context) Error() (string, error) {
 	return ctx.TaskName, ctx.err
 }
 
+// Get
 func (ctx *Context) Get(name string) any {
 	return ctx.params[name]
 }
 
+// Set
 func (ctx *Context) Set(name string, val any) {
 	ctx.params[name] = val
 }
