+++ mfs/mfs.go
@@ -2,35 +2,25 @@ package mfs
 
 import "errors"
 
-type TaskMethod struct {
-	TaskName   string // taskName
-	TaskMethod Method // TaskMethod
+// ITaskMethod abstract
+type ITaskMethod interface {
+	Name() string    // get the task name
+	Method(*Context) // task method
 }
 
-func (t *TaskMethod) Name() string {
-	return t.TaskName
-}
-
-func (t *TaskMethod) Runner(ctx *Context) {
-	t.TaskMethod(ctx)
-}
-
-func CreateMethod(name string, taskMethod func(ctx *Context)) TaskMethod {
-	return TaskMethod{
-		TaskName:   name,
-		TaskMethod: taskMethod,
-	}
-}
-
+// Method
 type Method func(*Context)
 
-type MethodChain []TaskMethod
+// MethodChain
+type MethodChain []ITaskMethod
 
+// Factory
 type Factory struct {
 	*Context
 	methodChain MethodChain // MethodChain
 }
 
+// NewFactory
 func NewFactory() *Factory {
 	f := &Factory{
 		methodChain: make(MethodChain, 0),
@@ -44,10 +34,16 @@ func NewFactory() *Factory {
 	return f
 }
 
-func (f *Factory) AddTaskMethod(taskMethods ...TaskMethod) {
+// AddTaskMethod as a write method area,
+// it is not recommended for developers to call this method when running the factory,
+// otherwise uncontrollable results may occur
+func (f *Factory) AddTaskMethod(taskMethods ...ITaskMethod) {
 	f.methodChain = append(f.methodChain, taskMethods...)
 }
 
+// DropTaskByName as a deletion method area,
+// it is not recommended for developers to call this method when running the factory,
+// otherwise uncontrollable results may occur
 func (f *Factory) DropTaskByName(name string) {
 	newMethodChain := make(MethodChain, 0)
 	for _, method := range f.methodChain {
@@ -59,7 +55,8 @@ func (f *Factory) DropTaskByName(name string) {
 	f.methodChain = newMethodChain
 }
 
-func (f *Factory) Run() {
+// Excute
+func (f *Factory) Excute() {
 	f.next = make(chan int)
 	go func() {
 		defer close(f.next)
@@ -77,8 +74,8 @@ func (f *Factory) Run() {
 					return
 				} else {
 					f.Now = ext
-					f.TaskName = f.methodChain[ext].TaskName
-					go f.methodChain[ext].TaskMethod(f.Context)
+					f.TaskName = f.methodChain[ext].Name()
+					go f.methodChain[ext].Method(f.Context)
 				}
 			}
 		}
